+++ themis-dashboard/src/app/(judge)/judge/cases/data/cases.ts
@@ -1,0 +1,71 @@
+
+export const cases = [
+    {
+      caseNo: "CASE22465",
+      title: "Irungii Khoikho Khamati",
+      accuracy: "80%",
+      date: "26th August 2023",
+      time: "22:45hrs",
+      status: "CLOSED CASE",
+    },
+    {
+      caseNo: "CASE545",
+      title: "Idi Amin",
+      accuracy: "80%",
+      date: "2 January 2017",
+      time: "13:45hrs",
+      status: "OPEN CASE",
+    },
+    {
+        caseNo: "CASE22475",
+        title: "Irungii Khoikho Khamati",
+        accuracy: "80%",
+        date: "26th August 2023",
+        time: "22:45hrs",
+        status: "CLOSED CASE",
+      },
+      {
+        caseNo: "CASE595",
+        title: "Idi Amin",
+        accuracy: "80%",
+        date: "2 January 2017",
+        time: "13:45hrs",
+        status: "OPEN CASE",
+      },
+      {
+        caseNo: "CASE22485",
+        title: "Irungii Khoikho Khamati",
+        accuracy: "80%",
+        date: "26th August 2023",
+        time: "22:45hrs",
+        status: "CLOSED CASE",
+      },
+      {
+        caseNo: "CASE586",
+        title: "Idi Amin",
+        accuracy: "80%",
+        date: "2 January 2017",
+        time: "13:45hrs",
+        status: "OPEN CASE",
+      },
+      {
+        caseNo: "CASE22455",
+        title: "Irungii Khoikho Khamati",
+        accuracy: "80%",
+        date: "26th August 2023",
+        time: "22:45hrs",
+        status: "CLOSED CASE",
+      },
+      {
+        caseNo: "CASE525",
+        title: "Idi Amin",
+        accuracy: "80%",
+        date: "2 January 2017",
+        time: "13:45hrs",
+        status: "OPEN CASE",
+      },
+   
+  ];
+  
+  export const getFirstThreeCases = () => cases.slice(0, 3);
+
